backend/requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
Flask-Cors==3.0.10
Pillow==9.0.0
pytesseract==0.3.8